+++ CHANGELOG.md
@@ -9,12 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Specify docker image when publishing a module
 - Improvements for AsyncConnector.
 - Improvements for Async Http workflow
 - Remove duplicated parts and make the code more uniform for async http workflow
 
 ### Fixed
 
+- Replace ulrllib.parse.urljoin by posixpath.join in AsyncConnector
 - Fix tests for async version of connector.
 
 ## 1.16.0 - 2024-10-16